added example_bullets component
