+++ components/benefit_cards.js
@@ -7,7 +7,7 @@ import Paper from "./paper";
 import { connect } from "react-redux";
 import NeedTag from "./need_tag";
 import { css } from "emotion";
-import RelatedBenefits from "./related_benefits_list";
+import BenefitExpansion from "./benefit_expansion";
 import BenefitCardHeader from "./benefit_card_header";
 import OneLiner from "./typography/one_liner";
 import Header from "./typography/header";
@@ -113,7 +113,8 @@ export class BenefitCard extends Component {
               ))}
             </div>
             {this.state.expanded ? (
-              <RelatedBenefits
+              <BenefitExpansion
+                className={padding}
                 benefit={benefit}
                 t={t}
                 store={this.props.store}
